Create password-management.md
This is the current password policy on MyAccount. Until we are able to influence desired changes to password policy on MyAccount, adding a different password policy as part of RA Wiki will likely cause more confusion and double standards, as most are relying or referencing MyAccount standards. Open to discussion.