refactor(index): bring back the index page
Reference architecture wiki should let you see the core RA technologies & traditions at-a-glance, so that it can be used as a tech radar & instructional repo. With the sections buried, and many concerns overlapping, its hard to know where to look.

I have left out the testing section as it has heavily strayed from the original index.