docs(react): revise for clarity, grammar, and formality (#197)
The reason why the Why section has changed to bullet points is that
the entire paragraph reduces to a list of items required when choosing React.
The point form bulleted structure is easier to understand for listed attributes in
technical writing.

In the What and How section, conversational colloquialisms are removed
in favour for a more formal style.

Links to more information on how React works is given rather than trying
to explain it all in this document.

There are some cases where quotation marks are used around a single word.
These are called "scare quotes" and are used for emphasis only when
quoting words someone else used. It usually implies that the author does not agree
with the use of the terms. Reference: https://www.grammarly.com/blog/quotation-marks-around-a-single-word/